streamlit/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pandas
pdfplumber
python-dotenv
